+++ lib/cuprum/built_in/identity_command.rb
@@ -1,11 +1,11 @@
 require 'cuprum/built_in'
-require 'cuprum/function'
+require 'cuprum/command'
 
 module Cuprum::BuiltIn
   # A predefined function that returns the value or result it was called with.
   #
   # @example With a value.
-  #   result = IdentityFunction.new.call('custom value')
+  #   result = IdentityCommand.new.call('custom value')
   #   result.value
   #   #=> 'custom value'
   #   result.success?
@@ -14,14 +14,14 @@ module Cuprum::BuiltIn
   # @example With a result.
   #   errors = ['errors.messages.unknown']
   #   value  = Cuprum::Result.new('result value', :errors => errors)
-  #   result = IdentityFunction.new.call(value)
+  #   result = IdentityCommand.new.call(value)
   #   result.value
   #   #=> 'result value'
   #   result.success?
   #   #=> false
   #   result.errors
   #   #=> ['errors.messages.unknown']
-  class IdentityFunction < Cuprum::Function
+  class IdentityCommand < Cuprum::Command
     private
 
     def process value = nil