+++ lib/cuprum/built_in/null_command.rb
@@ -1,16 +1,16 @@
 require 'cuprum/built_in'
-require 'cuprum/function'
+require 'cuprum/command'
 
 module Cuprum::BuiltIn
-  # A predefined function that does nothing when called.
+  # A predefined command that does nothing when called.
   #
   # @example
-  #   result = NullFunction.new.call
+  #   result = NullCommand.new.call
   #   result.value
   #   #=> nil
   #   result.success?
   #   #=> true
-  class NullFunction < Cuprum::Function
+  class NullCommand < Cuprum::Command
     private
 
     def process *_args; end