+++ spec/cuprum/basic_command_spec.rb
@@ -1,16 +1,14 @@
-require 'cuprum/operation'
+require 'cuprum/basic_command'
 
 require 'support/examples/command_examples'
-require 'support/examples/operation_examples'
 
-RSpec.describe Cuprum::Operation do
+RSpec.describe Cuprum::BasicCommand do
   include Spec::Examples::CommandExamples
-  include Spec::Examples::OperationExamples
 
   subject(:instance) { described_class.new }
 
   let(:implementation) { ->() {} }
-  let(:result_class)   { described_class }
+  let(:result_class)   { Cuprum::Result }
 
   describe '::new' do
     it { expect(described_class).to be_constructible.with(0).arguments }
@@ -19,6 +17,4 @@ RSpec.describe Cuprum::Operation do
   include_examples 'should implement the Command methods'
 
   include_examples 'should implement the Command methods for any implementation'
-
-  include_examples 'should implement the Operation methods'
 end # describe