+++ spec/cuprum/built_in/identity_operation_spec.rb
@@ -1,9 +1,10 @@
 require 'cuprum/built_in/identity_operation'
-require 'cuprum/function_examples'
-require 'cuprum/operation_examples'
+
+require 'support/examples/command_examples'
+require 'support/examples/operation_examples'
 
 RSpec.describe Cuprum::BuiltIn::IdentityOperation do
-  include Spec::Examples::FunctionExamples
+  include Spec::Examples::CommandExamples
   include Spec::Examples::OperationExamples
 
   subject(:instance) { described_class.new }
@@ -12,7 +13,7 @@ RSpec.describe Cuprum::BuiltIn::IdentityOperation do
   let(:errors) { ['errors.messages.unknown'] }
   let(:result) { Cuprum::Result.new(value, :errors => errors) }
 
-  include_examples 'should implement the Function methods'
+  include_examples 'should implement the Command methods'
 
   include_examples 'should implement the Operation methods'