+++ spec/cuprum/command_spec.rb
@@ -1,16 +1,16 @@
-require 'cuprum/operation'
+require 'cuprum/command'
 
+require 'support/examples/chaining_examples'
 require 'support/examples/command_examples'
-require 'support/examples/operation_examples'
 
-RSpec.describe Cuprum::Operation do
+RSpec.describe Cuprum::Command do
+  include Spec::Examples::ChainingExamples
   include Spec::Examples::CommandExamples
-  include Spec::Examples::OperationExamples
 
   subject(:instance) { described_class.new }
 
   let(:implementation) { ->() {} }
-  let(:result_class)   { described_class }
+  let(:result_class)   { Cuprum::Result }
 
   describe '::new' do
     it { expect(described_class).to be_constructible.with(0).arguments }
@@ -20,5 +20,9 @@ RSpec.describe Cuprum::Operation do
 
   include_examples 'should implement the Command methods for any implementation'
 
-  include_examples 'should implement the Operation methods'
+  include_examples 'should implement the Command chaining methods'
+
+  describe '#errors' do
+    include_examples 'should have private reader', :errors
+  end # describe
 end # describe